Managed-Identity-App/requirements.txt
azure-core==1.35.0
azure-identity==1.24.0
azure-keyvault-secrets==4.10.0
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
Flask==3.1.1
gunicorn==23.0.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msal==1.33.0
msal-extensions==1.3.1
packaging==25.0
pycparser==2.22
PyJWT==2.10.1
requests==2.32.4
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
